rename sub_pipe -> subprocess_pipes
